Re-organize the api routes.
