+++ mash/services/api/routes/accounts/ec2.py
@@ -20,33 +20,33 @@ import json
 from flask import jsonify, request, make_response
 from flask_restplus import Namespace, Resource
 
-from mash.services.api.accounts import (
+from mash.services.api.routes.accounts import (
     account_response,
     validation_error_response,
     delete_account_request
 )
-from mash.services.api.schema.accounts.azure import add_account_azure
-from mash.services.api.utils import publish
+from mash.services.api.schema.accounts.ec2 import add_account_ec2
+from mash.services.api.routes.utils import publish
 
 api = Namespace(
-    'Azure Accounts',
-    description='Azure account related operations'
+    'EC2 Accounts',
+    description='EC2 account related operations'
 )
-azure_account = api.schema_model('azure_account', add_account_azure)
+ec2_account = api.schema_model('ec2_account', add_account_ec2)
 
 
 @api.route('/')
 @api.response(400, 'Validation error', validation_error_response)
-class AzureAccountList(Resource):
-    @api.doc('create_azure_account')
-    @api.expect(azure_account)
-    @api.response(201, 'Azure account created', account_response)
+class EC2AccountList(Resource):
+    @api.doc('create_ec2_account')
+    @api.expect(ec2_account)
+    @api.response(201, 'EC2 account created', account_response)
     def post(self):
         """
-        Create a new Azure account.
+        Create a new EC2 account.
         """
         data = json.loads(request.data.decode())
-        data['cloud'] = 'azure'
+        data['cloud'] = 'ec2'
 
         publish(
             'jobcreator', 'add_account', json.dumps(data, sort_keys=True)
@@ -56,16 +56,16 @@ class AzureAccountList(Resource):
 
 @api.route('/<int:id>')
 @api.response(400, 'Validation error', validation_error_response)
-class AzureAccount(Resource):
-    @api.doc('delete_azure_account')
+class EC2Account(Resource):
+    @api.doc('delete_ec2_account')
     @api.expect(delete_account_request)
-    @api.response(200, 'Azure account deleted', account_response)
+    @api.response(200, 'EC2 account deleted', account_response)
     def delete(self, id):
         """
-        Delete Azure account matching id.
+        Delete EC2 account matching id.
         """
         data = json.loads(request.data.decode())
-        data['cloud'] = 'azure'
+        data['cloud'] = 'ec2'
 
         publish(
             'jobcreator', 'delete_account', json.dumps(data, sort_keys=True)
