+++ test/unit/services/api/api_test.py
@@ -17,8 +17,8 @@ def test_client():
     ctx.pop()
 
 
-@patch('mash.services.api.utils.BaseConfig')
-@patch('mash.services.api.utils.Connection')
+@patch('mash.services.api.routes.utils.BaseConfig')
+@patch('mash.services.api.routes.utils.Connection')
 def test_api_add_account_ec2(mock_connection, mock_config, test_client):
     channel = MagicMock()
     connection = MagicMock()
@@ -63,8 +63,8 @@ def test_api_add_account_ec2(mock_connection, mock_config, test_client):
     assert response.data == b'{"name":"test"}\n'
 
 
-@patch('mash.services.api.utils.BaseConfig')
-@patch('mash.services.api.utils.Connection')
+@patch('mash.services.api.routes.utils.BaseConfig')
+@patch('mash.services.api.routes.utils.Connection')
 def test_api_add_account_gce(mock_connection, mock_config, test_client):
     channel = MagicMock()
     connection = MagicMock()
@@ -117,8 +117,8 @@ def test_api_add_account_gce(mock_connection, mock_config, test_client):
     assert response.data == b'{"name":"test"}\n'
 
 
-@patch('mash.services.api.utils.BaseConfig')
-@patch('mash.services.api.utils.Connection')
+@patch('mash.services.api.routes.utils.BaseConfig')
+@patch('mash.services.api.routes.utils.Connection')
 def test_api_add_account_azure(mock_connection, mock_config, test_client):
     channel = MagicMock()
     connection = MagicMock()
@@ -176,8 +176,8 @@ def test_api_add_account_azure(mock_connection, mock_config, test_client):
     assert response.data == b'{"name":"test"}\n'
 
 
-@patch('mash.services.api.utils.BaseConfig')
-@patch('mash.services.api.utils.Connection')
+@patch('mash.services.api.routes.utils.BaseConfig')
+@patch('mash.services.api.routes.utils.Connection')
 def test_api_delete_account_ec2(mock_connection, mock_config, test_client):
     channel = MagicMock()
     connection = MagicMock()
@@ -215,8 +215,8 @@ def test_api_delete_account_ec2(mock_connection, mock_config, test_client):
     assert response.data == b'{"name":"test"}\n'
 
 
-@patch('mash.services.api.utils.BaseConfig')
-@patch('mash.services.api.utils.Connection')
+@patch('mash.services.api.routes.utils.BaseConfig')
+@patch('mash.services.api.routes.utils.Connection')
 def test_api_delete_account_gce(mock_connection, mock_config, test_client):
     channel = MagicMock()
     connection = MagicMock()
@@ -254,8 +254,8 @@ def test_api_delete_account_gce(mock_connection, mock_config, test_client):
     assert response.data == b'{"name":"test"}\n'
 
 
-@patch('mash.services.api.utils.BaseConfig')
-@patch('mash.services.api.utils.Connection')
+@patch('mash.services.api.routes.utils.BaseConfig')
+@patch('mash.services.api.routes.utils.Connection')
 def test_api_delete_account_azure(mock_connection, mock_config, test_client):
     channel = MagicMock()
     connection = MagicMock()
@@ -293,9 +293,9 @@ def test_api_delete_account_azure(mock_connection, mock_config, test_client):
     assert response.data == b'{"name":"test"}\n'
 
 
-@patch('mash.services.api.utils.BaseConfig')
-@patch('mash.services.api.jobs.uuid')
-@patch('mash.services.api.utils.Connection')
+@patch('mash.services.api.routes.utils.BaseConfig')
+@patch('mash.services.api.routes.jobs.uuid')
+@patch('mash.services.api.routes.utils.Connection')
 def test_api_add_job_ec2(mock_connection, mock_uuid, mock_config, test_client):
     channel = MagicMock()
     connection = MagicMock()
@@ -339,9 +339,9 @@ def test_api_add_job_ec2(mock_connection, mock_uuid, mock_config, test_client):
         b'{"job_id": "12345678-1234-1234-1234-123456789012"}'
 
 
-@patch('mash.services.api.utils.BaseConfig')
-@patch('mash.services.api.jobs.uuid')
-@patch('mash.services.api.utils.Connection')
+@patch('mash.services.api.routes.utils.BaseConfig')
+@patch('mash.services.api.routes.jobs.uuid')
+@patch('mash.services.api.routes.utils.Connection')
 def test_api_add_job_gce(mock_connection, mock_uuid, mock_config, test_client):
     channel = MagicMock()
     connection = MagicMock()
@@ -385,9 +385,9 @@ def test_api_add_job_gce(mock_connection, mock_uuid, mock_config, test_client):
         b'{"job_id": "12345678-1234-1234-1234-123456789012"}'
 
 
-@patch('mash.services.api.utils.BaseConfig')
-@patch('mash.services.api.jobs.uuid')
-@patch('mash.services.api.utils.Connection')
+@patch('mash.services.api.routes.utils.BaseConfig')
+@patch('mash.services.api.routes.jobs.uuid')
+@patch('mash.services.api.routes.utils.Connection')
 def test_api_add_job_azure(mock_connection, mock_uuid, mock_config, test_client):
     channel = MagicMock()
     connection = MagicMock()
@@ -431,8 +431,8 @@ def test_api_add_job_azure(mock_connection, mock_uuid, mock_config, test_client)
         b'{"job_id": "12345678-1234-1234-1234-123456789012"}'
 
 
-@patch('mash.services.api.utils.BaseConfig')
-@patch('mash.services.api.utils.Connection')
+@patch('mash.services.api.routes.utils.BaseConfig')
+@patch('mash.services.api.routes.utils.Connection')
 def test_api_delete_job(mock_connection, mock_config, test_client):
     channel = MagicMock()
     connection = MagicMock()
